Moved EditQuestion and EditAnswer to subComponents of QuestionAnswerViewer
